requirements/base.txt
Django==1.6.5
South==0.8.4
Unipath==1.0
psycopg2==2.5.2
python-social-auth==0.1.22
djrill==1.2.0 
django-braces==1.4.0 

